Worked on Parser and Node functions
Added node functions where he asked and got the basic structure of the
exp down.  I had trouble with the rest grammar and not quite sure how to
work that.  We will implement the tree at a later time after he goes
over it in class.  Any extras todos that i made were marked with GRANT
above it.