Parse Tree Update
Worked on the parse Tree implementation. I changed the grammar to fix
left recursion and common left factors.  I added nodes to all grammar
rules.  I think I completed it but I haven't tested anything, because
its hard to do so.  We still have to work on the special class and
implement a function in the cons class.